third_party/WebKit/LayoutTests/fast/alignment/parse-align-self-expected.txt

Test that setting and getting align-self works as expected

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Test getting align-self set through CSS
PASS getComputedStyle(alignSelfBaseline, '').getPropertyValue('align-self') is 'baseline'
PASS getComputedStyle(alignSelfLastBaseline, '').getPropertyValue('align-self') is 'last-baseline'
PASS getComputedStyle(alignSelfStretch, '').getPropertyValue('align-self') is 'stretch'
PASS getComputedStyle(alignSelfStart, '').getPropertyValue('align-self') is 'start'
PASS getComputedStyle(alignSelfEnd, '').getPropertyValue('align-self') is 'end'
PASS getComputedStyle(alignSelfCenter, '').getPropertyValue('align-self') is 'center'
PASS getComputedStyle(alignSelfSelfEnd, '').getPropertyValue('align-self') is 'self-end'
PASS getComputedStyle(alignSelfSelfStart, '').getPropertyValue('align-self') is 'self-start'
PASS getComputedStyle(alignSelfLeft, '').getPropertyValue('align-self') is 'left'
PASS getComputedStyle(alignSelfRight, '').getPropertyValue('align-self') is 'right'
PASS getComputedStyle(alignSelfFlexStart, '').getPropertyValue('align-self') is 'flex-start'
PASS getComputedStyle(alignSelfFlexEnd, '').getPropertyValue('align-self') is 'flex-end'
PASS getComputedStyle(alignSelfEndUnsafe, '').getPropertyValue('align-self') is 'end unsafe'
PASS getComputedStyle(alignSelfCenterUnsafe, '').getPropertyValue('align-self') is 'center unsafe'
PASS getComputedStyle(alignSelfSelfEndSafe, '').getPropertyValue('align-self') is 'self-end safe'
PASS getComputedStyle(alignSelfSelfStartSafe, '').getPropertyValue('align-self') is 'self-start safe'
PASS getComputedStyle(alignSelfRightSafe, '').getPropertyValue('align-self') is 'right safe'
PASS getComputedStyle(alignSelfLeftUnsafe, '').getPropertyValue('align-self') is 'left unsafe'
PASS getComputedStyle(alignSelfFlexStartUnsafe, '').getPropertyValue('align-self') is 'flex-start unsafe'
PASS getComputedStyle(alignSelfFlexEndSafe, '').getPropertyValue('align-self') is 'flex-end safe'

Test initial value of align-self through JS
PASS getComputedStyle(element, '').getPropertyValue('align-self') is 'normal'

Test getting and setting align-self through JS
PASS element.style.alignSelf is "center"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "center"
PASS element.style.alignSelf is "start unsafe"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "start unsafe"
PASS element.style.alignSelf is "flex-end safe"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "flex-end safe"
PASS element.style.alignSelf is "right"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "right"
PASS element.style.alignSelf is "center"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "center"
PASS element.style.alignSelf is "self-start"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "self-start"
PASS element.style.alignSelf is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is "self-end"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "self-end"

Test 'auto' value resolution for the root node
PASS element.style.alignSelf is "auto"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test bad combinations of align-self
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"
PASS element.style.alignSelf is ""
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test the value 'initial'
PASS element.style.alignSelf is "center"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "center"
PASS element.style.alignSelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test the value 'initial' for grid containers
PASS element.style.alignSelf is "left safe"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "left safe"
PASS element.style.alignSelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test the value 'initial' for flex containers
PASS element.style.alignSelf is "right unsafe"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "right unsafe"
PASS element.style.alignSelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test the value 'initial' for positioned elements
PASS element.style.alignSelf is "left"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "left"
PASS element.style.alignSelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test the value 'initial' for positioned elements in grid containers
PASS element.style.alignSelf is "right"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "right"
PASS element.style.alignSelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test the value 'initial' for positioned elements in grid containers
PASS element.style.alignSelf is "end"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "end"
PASS element.style.alignSelf is "initial"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "normal"

Test the value 'inherit'
PASS element.style.alignSelf is "end"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "end"
PASS element.style.alignSelf is "inherit"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "end"
PASS element.style.alignSelf is "left safe"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "left safe"
PASS element.style.alignSelf is "inherit"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "left safe"
PASS element.style.alignSelf is "center unsafe"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "center unsafe"
PASS element.style.alignSelf is "inherit"
PASS window.getComputedStyle(element, '').getPropertyValue('align-self') is "center unsafe"
PASS successfullyParsed is true

TEST COMPLETE
